requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pdfplumber==0.11.4
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
schedule==1.2.2
redis==5.2.0
pandas==2.2.3
